class_7/class_code/human-in-the-loop/requirements.txt
langchain
openai
langsmith
streamlit
python-dotenv
fastapi
uvicorn
requests 